requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
boto3==1.14.6
botocore==1.17.6
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coverage==5.1
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap-modal-forms==1.5.0
django-ckeditor==5.9.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-js-asset==1.2.2
django-storages==1.9.1
django-tinymce==3.0.2
django-widget-tweaks==1.4.8
docutils==0.15.2
flake8==3.8.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
stripe==2.48.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
zipp==3.1.0
